feat: Add Esigner class
