+++ toolbelt/exceptions.py
@@ -6,7 +6,7 @@ class ResponseError(ClientError):
     pass
 
 
-class PlanetError(Exception):
+class CliError(Exception):
     def __init__(self, cmd: str, msg: str):
         super().__init__(msg)
 
@@ -14,6 +14,14 @@ class PlanetError(Exception):
 
     def __str__(self):
         return f"Command: {self.cmd}, Message: {super().__str__()}"
+
+
+class PlanetError(CliError):
+    pass
+
+
+class EsignerError(CliError):
+    pass
 
 
 class TagNotFoundError(Exception):
